requirements.txt
jina==3.1.1
Pillow==9.0.1
pytesseract==0.3.9
